solved #1159
